typeinfo数据落盘
# message auto-generated for no-merge-commit merge:
merge ctrlc0925 into master

typeinfo数据落盘

Created-by: msprofDeveloper

Match-id-b4640a6df0f24b6ba57d51a7d58125cb593d6630

+++ collector/dvvp/transport/hash_data.h
@@ -56,7 +56,7 @@ private:
     std::mutex initMutex_;
     std::mutex hashMutex_;
     bool isStarted_ = false;
-    int32_t readIndex_ = 0;
+    size_t readIndex_ = 0;
     std::vector<std::pair<uint64_t, std::string>> hashVector_;
     std::unordered_map<std::string, uint64_t> hashInfoMap_;   // <hashInfo, hashId>
     std::unordered_map<uint64_t, std::string> hashIdMap_;     // <hashId, hashInfo>
